Move Nav Bar Buttons into individual file, initialize Nav Bar, style page with some css
